+++ playground/app.js
@@ -6,6 +6,7 @@ const {
   filterOutDomain,
   getTranslator,
   mongoResultsToDataset,
+  PipelineDereferencer,
   servicePluginFactory,
   registerModule,
 } = vqb;
@@ -39,7 +40,10 @@ function annotateDataset(dataset, typeAnnotations) {
     name: hdr.name,
     type: typeAnnotations ? mongoToVQBType(typeAnnotations[hdr.name]) : undefined,
   }));
-  return { ...dataset, headers: typedHeaders };
+  return {
+    ...dataset,
+    headers: typedHeaders
+  };
 }
 
 
@@ -66,7 +70,11 @@ function autocastDataset(dataset) {
     }
     newData.push(newRow);
   }
-  return { ...dataset, data: newData, headers: dataset.headers };
+  return {
+    ...dataset,
+    data: newData,
+    headers: dataset.headers
+  };
 }
 
 
@@ -77,9 +85,15 @@ class MongoService {
   }
 
   async executePipeline(_store, pipeline, limit, offset = 0) {
-    const { domain, pipeline: subpipeline } = filterOutDomain(pipeline);
+    const {
+      domain,
+      pipeline: subpipeline
+    } = filterOutDomain(pipeline);
     const query = mongo36translator.translate(subpipeline);
-    const { isResponseOk, responseContent } = await this.executeQuery(query, domain, limit, offset);
+    const {
+      isResponseOk,
+      responseContent
+    } = await this.executeQuery(query, domain, limit, offset);
 
     if (isResponseOk) {
       const [{
@@ -140,15 +154,20 @@ class MongoService {
 
 const mongoservice = new MongoService();
 const mongoBackendPlugin = servicePluginFactory(mongoservice);
-const initialPipeline = [
-  {
-    name: 'domain',
-    domain: 'test-collection',
+const initialPipeline = [{
+  name: 'domain',
+  domain: 'test-collection',
+},
+{
+  name: 'filter',
+  condition: {
+    and: [{
+      column: 'Groups',
+      value: '<%= groupname %>',
+      operator: 'eq'
+    }]
   },
-  {
-    name: 'filter',
-    condition: { and: [{ column: 'Groups', value: '<%= groupname %>', operator: 'eq' }] },
-  },
+},
 ];
 
 async function setupInitialData(store, domain = null) {
@@ -168,10 +187,15 @@ async function setupInitialData(store, domain = null) {
     );
     if (response.error) {
       store.commit(VQBnamespace('setBackendError'), {
-        backendError: { type: 'error', error: response.error },
+        backendError: {
+          type: 'error',
+          error: response.error
+        },
       });
     } else {
-      store.commit(VQBnamespace('setDataset'), { dataset: response.data });
+      store.commit(VQBnamespace('setDataset'), {
+        dataset: response.data
+      });
     }
   }
 }
@@ -199,6 +223,40 @@ async function buildVueApp() {
     created: function() {
       registerModule(this.$store, {
         pipeline: initialPipeline,
+        pipelines: {
+          pipeline1: [{
+            name: 'domain',
+            domain: 'test-collection',
+          },
+          {
+            name: 'replace',
+            search_column: 'Label',
+            to_replace: [
+              ['Label 1', 'Label 6'],
+              ['Label 2', 'Label 7'],
+              ['Label 3', 'Label 8'],
+              ['Label 4', 'Label 9'],
+              ['Label 5', 'Label 10'],
+            ]
+          },
+          ],
+          pipeline2: [{
+            name: 'domain',
+            domain: 'test-collection',
+          },
+          {
+            name: 'replace',
+            search_column: 'Label',
+            to_replace: [
+              ['Label 1', 'Label 11'],
+              ['Label 2', 'Label 12'],
+              ['Label 3', 'Label 13'],
+              ['Label 4', 'Label 14'],
+              ['Label 5', 'Label 15'],
+            ]
+          },
+          ],
+        },
         currentDomain: 'test-collection',
         // use lodash interpolate
         interpolateFunc: (value, context) => _.template(value)(context),
@@ -212,9 +270,13 @@ async function buildVueApp() {
     },
     computed: {
       code: function() {
-        const query = mongo36translator.translate(
-          this.$store.getters[VQBnamespace('activePipeline')],
-        );
+        let activePipeline = this.$store.getters[VQBnamespace('activePipeline')]
+        const pipelines = this.$store.getters[VQBnamespace('pipelines')]
+        if (pipelines) {
+          const dereferencer = new PipelineDereferencer(pipelines);
+          activePipeline = dereferencer.dereference(activePipeline);
+        }
+        const query = mongo36translator.translate(activePipeline);
         return JSON.stringify(query, null, 2);
       },
       thereIsABackendError: function() {
@@ -254,7 +316,9 @@ async function buildVueApp() {
         this.draggedover = false;
         event.preventDefault();
         // For the moment, only take one file and we should also test event.target
-        const { collection: domain } = await mongoservice.loadCSV(event.dataTransfer.files[0]);
+        const {
+          collection: domain
+        } = await mongoservice.loadCSV(event.dataTransfer.files[0]);
         await setupInitialData(store, domain);
         event.target.value = null;
       },